requirements.txt
Flask
sentence-transformers
faiss-cpu
openai==0.28
python-dotenv
